// types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      cargas: {
        Row: {
          dador_id: string
          email: string | null
          fecha_creacion: string
          fechacarga: string
          fechadescarga: string
          formadepago_id: string
          id: string
          material_id: string
          otropagopor: string | null
          pagopor: string
          peso: string
          presentacion_id: string
          puntoreferencia: string | null
          telefonodador: string
          tipo_equipo: string | null
          ubicacionfinal_id: string
          ubicacioninicial_id: string
          valorviaje: string
        }
        Insert: {
          dador_id: string
          email?: string | null
          fecha_creacion?: string
          fechacarga: string
          fechadescarga: string
          formadepago_id: string
          id?: string
          material_id: string
          otropagopor?: string | null
          pagopor: string
          peso: string
          presentacion_id: string
          puntoreferencia?: string | null
          telefonodador: string
          tipo_equipo?: string | null
          ubicacionfinal_id: string
          ubicacioninicial_id: string
          valorviaje: string
        }
        Update: {
          dador_id?: string
          email?: string | null
          fecha_creacion?: string
          fechacarga?: string
          fechadescarga?: string
          formadepago_id?: string
          id?: string
          material_id?: string
          otropagopor?: string | null
          pagopor?: string
          peso?: string
          presentacion_id?: string
          puntoreferencia?: string | null
          telefonodador?: string
          tipo_equipo?: string | null
          ubicacionfinal_id?: string
          ubicacioninicial_id?: string
          valorviaje?: string
        }
        Relationships: [
          {
            foreignKeyName: "cargas_dador_id_fkey"
            columns: ["dador_id"]
            isOneToOne: false
            referencedRelation: "usuarios"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cargas_formadepago_id_fkey"
            columns: ["formadepago_id"]
            isOneToOne: false
            referencedRelation: "formadepago"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cargas_material_id_fkey"
            columns: ["material_id"]
            isOneToOne: false
            referencedRelation: "material"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cargas_presentacion_id_fkey"
            columns: ["presentacion_id"]
            isOneToOne: false
            referencedRelation: "presentacion"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cargas_ubicacionfinal_id_fkey"
            columns: ["ubicacionfinal_id"]
            isOneToOne: false
            referencedRelation: "ubicaciones"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "cargas_ubicacioninicial_id_fkey"
            columns: ["ubicacioninicial_id"]
            isOneToOne: false
            referencedRelation: "ubicaciones"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      ubicaciones: {
        Row: {
          ciudad_id: number
          direccion: string
          id: string
          lat: number
          lng: number
          provincia_id: number
        }
        Insert: {
          ciudad_id: number
          direccion: string
          id?: string
          lat: number
          lng: number
          provincia_id: number
        }
        Update: {
          ciudad_id?: number
          direccion?: string
          id?: string
          lat?: number
          lng?: number
          provincia_id?: number
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}